requirements.txt
jira==3.8.0
python-dateutil==2.9.0.post0
numpy==2.0.1
sounddevice==0.4.7
